Function for summarizing the uncertainty propagation's results in the form of a unique CDF.

Function for summarizing the uncertainty propagation's results in the form of a unique CDF via the weighting average approach of Dubois and Guyonnet (2011).

summary_1cdf(Z0, aversion = 0.5)

Arguments

Z0

Output of the uncertainty propagation function PROPAG()

aversion

Weight value representing the decision-maker risk aversion i.e. the balance between the lower and upper CDFs. by default, alpha=0.5.

Value

Vector of the same size as the number of columns of Z0.

Details

Details of the theory and the example in Dubois & Guyonnet (2011) Available at: https://hal-brgm.archives-ouvertes.fr/file/index/docid/578821/filename/Uncertainties_RA_09_l_dg.pdf

References

Dubois D., Guyonnet D. 2011. Risk-informed decision-making under epistemic uncertainty. International Journal of General Systems, 40(2), 145-167.

See also

Examples

if (FALSE) {
#################################################
#### EXAMPLE 1 of Dubois & Guyonnet (2011)
#### Probability and Possibility distributions
#################################################

#### Model function
FUN = function(X){
UER=X[1]
EF=X[2]
I=X[3]
C=X[4]
ED=X[5]
return(UER*I*C*EF*ED/(70*70*365))
}

ninput = 5 #Number of input parameters
input = vector(mode="list", length=ninput) # Initialisation

input[[1]] = create_input(
    name="UER",
    type="possi",
    distr="triangle",
    param=c(2.e-2, 5.7e-2, 1.e-1),
    monoton="incr"
    )
input[[2]] = create_input(
    name="EF",
    type="possi",
    distr="triangle",
    param=c(200,250,350),
    monoton="incr"
    )
input[[3]] = create_input(
    name="I",
    type="possi",
    distr="triangle",
    param=c(1,1.5,2.5),
    monoton="incr"
    )
input[[4]] = create_input(
    name="C",
    type="proba",
    distr="triangle",
    param=c(5e-3,20e-3,10e-3)
    )
input[[5]] = create_input(
    name="ED",
    type="proba",
    distr="triangle",
    param=c(10,50,30)
    )

####CREATION OF THE DISTRIBUTIONS ASSOCIATED TO THE PARAMETERS
input = create_distr(input)

####VISU INPUT
plot_input(input)

#################################################
#### PROPAGATION

#OPTIMZATION CHOICES
choice_opt = NULL #no optimization needed
param_opt = NULL

#PROPAGATION RUN
Z0_IRS = propag(N=1000, input, FUN, choice_opt, param_opt, mode="IRS")

#################################################
#### POST-PROCESSING

# VISU - PROPAGATION
plot_cdf(Z0_IRS, xlab="Z", ylab="CDF", main="EX 1", lwd=1.5)

# One CDF with risk aversion of 1/3
Z = summary_1cdf(Z0_IRS, aversion=1/3)
lines(ecdf(Z),col=5,lwd=1.5)

}