Function to perform pinching.

Function to pinch a imprecise variable to a fixed value following Ferson and Tucker (2006) and perform the propagation. Note that it only handles imprecise parametes. In this case of an imprecise probability distributions, only the imprecise parameters are handled.

pinching_fun(which, value, N, input, FUN, 
       choice_opt = "L-BFGS-B", param_opt = NULL, 
       corr = 0.01, NL = 10, mode = "IRS")

Arguments

which

Integer to specify the rank of the input variable as specifed in CREATE_INPUT(), i.e. parameter 1, 2,..., etc.

value

Scalar value to which the imprecise variable is pinched.

N

Integer corresponding to the number of random samples.

input

List of inputs as provided by create_input().

FUN

Model assessment function.

choice_opt

Option of constrainted optimization algorithm, see PROPAG.

param_opt

Parameters necessary for conducting the optimization algorithm, see PROPAG.

corr

Tolerance to avoid empty alpha-cuts. By default, corr=0.01.

NL

Integer to specify the number of alpha-cuts needed for hybrid propagation described by Baudrit et al. (2006). By default, NL=10.

mode

String to specify the mode of propagation: "IRS" (Baudrit et al. 2007) or "HYBRID" (Baudrit et al. 2006), see PROPAG.

References

Baudrit, C., Dubois, D., & Guyonnet, D. 2006. Joint propagation and exploitation of probabilistic and possibilistic information in risk assessment. IEEE transactions on fuzzy systems, 14(5), 593-608.

Baudrit, C., Guyonnet, D., Dubois, D. 2007. Joint propagation of variability and partial ignorance in a groundwater risk assessment. Journal of Contaminant Hydrology, 93: 72-84.

Ferson, S., & Tucker, W. T. (2006). Sensitivity analysis using probability bounding. Reliability Engineering & System Safety, 91(10), 1435-1442.

See also

Examples

#See example described for \emph{sensi_pinching}.